database-dimension/pyproject.toml
[build-system]
requires = ["uv_build>=0.6,<0.7"]
build-backend = "uv_build"

[project]
name = "database-dimension"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.10.0",
    "dev-pytopia @ git+https://github.com/CodeThatCodes/python-packages@main#subdirectory=dev-pytopia",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.uv.sources]
dev-pytopia = { path = "../dev-pytopia", editable = true }

[project.scripts]
database-dimension = "database_dimension.terminal_functionality:terminal_entrypoint"

[tool.uv.build-backend]
module-root = ""

[tool.ruff]
line-length = 115
src = ["database_dimension"]

[tool.ruff.lint]
select = ["F", "E", "I"]
ignore = ["E501", "E731"]
per-file-ignores = { "__init__.py" = ["F401"] }

[tool.ruff.lint.isort]
force-sort-within-sections = true
combine-as-imports = true
split-on-trailing-comma = true
relative-imports-order = "closest-to-furthest"
known-first-party = [
    "dev_pytopia",
]
